Adjust the directory structure & add some modules
