+++ GhUser/__init__.py
@@ -98,48 +98,3 @@ def GetUserFollowersList(User, DownloadAccountPicture=False):
             return(FollowersList)
         except:
             return(1)
-
-# 将用户数据导入到CSV中
-def GetUserDateToCSV(User):
-    requests.packages.urllib3.disable_warnings(InsecureRequestWarning) #禁用SSL证书验证警告
-    ApiURL = "https://api.github.com/users/"+User
-    InitJSON = requests.get(url=ApiURL, verify=False).text
-    LoadJSON = json.loads(InitJSON, strict=False)
-    try:
-        if LoadJSON["message"] == "Not Found":
-            return(1)
-    except:
-        pass
-    try:
-        with open("User.csv", "w", newline='', encoding="utf-8") as file:
-            writer = csv.writer(file)
-            writer.writerow(["UserName", "User", "Avatar","HomePage", "Company", "Blog", "Location", "Email", "Bio", "TwitterUsername", "Public Repos", "Public Gists", "Follower(s)", "Following", "Registration"])
-            if bool(LoadJSON["company"]) == False:
-                Company = "NO DATE"
-            else:
-                Company = LoadJSON["company"]
-            if bool(LoadJSON["blog"]) == False:
-                Blog = "NO DATE"
-            else:
-                Blog = LoadJSON["blog"]
-            if bool(LoadJSON["location"]) == False:
-                Location = "NO DATE"
-            else:
-                Location = LoadJSON["location"]
-            if bool(LoadJSON["email"]) == False:
-                Email = "NO DATE"
-            else:
-                Email = bool(LoadJSON["email"])
-            if bool(LoadJSON["bio"]) == False:
-                BIO = "NO DATE"
-            else:
-                BIO = LoadJSON["bio"]
-            if bool(LoadJSON["twitter_username"]) == False:
-                TwitterUsername = "NO DATE"
-            else:
-                TwitterUsername = LoadJSON["twitter_username"]
-            RegistrationTime = calendar.month_name[pd.to_datetime(LoadJSON["created_at"]).month]+" "+str(pd.to_datetime(LoadJSON["created_at"]).day)+", "+str(pd.to_datetime(LoadJSON["created_at"]).year)
-            writer.writerow([LoadJSON["name"], LoadJSON["login"], LoadJSON["avatar_url"], LoadJSON["html_url"], Company, Blog, Location, Email, BIO, TwitterUsername, str(LoadJSON["public_repos"]), str(LoadJSON["public_gists"]), str(LoadJSON["followers"]), str(LoadJSON["following"]), RegistrationTime])
-            return(0)
-    except:
-        return(1)
